Roll our own iOS application delegates
