+++ SampleGame.iOS/AppDelegate.cs
@@ -1,0 +1,15 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using Foundation;
+using osu.Framework;
+using osu.Framework.iOS;
+
+namespace SampleGame.iOS
+{
+    [Register("AppDelegate")]
+    public class AppDelegate : GameApplicationDelegate
+    {
+        protected override Game CreateGame() => new SampleGameGame();
+    }
+}
