+++ osu.Framework.Tests.iOS/AppDelegate.cs
@@ -1,0 +1,14 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using Foundation;
+using osu.Framework.iOS;
+
+namespace osu.Framework.Tests
+{
+    [Register("AppDelegate")]
+    public class AppDelegate : GameApplicationDelegate
+    {
+        protected override Game CreateGame() => new VisualTestGame();
+    }
+}
