+++ osu.Framework.Tests.iOS/Main.cs
@@ -1,0 +1,7 @@
+﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using osu.Framework.Tests;
+using UIKit;
+
+UIApplication.Main(args, null, typeof(AppDelegate));
